# [Java] Count digits in a number without using branches.