--- a/agrona/src/main/java/org/agrona/ExpandableArrayBuffer.java
+++ b/agrona/src/main/java/org/agrona/ExpandableArrayBuffer.java
@@ -22,10 +22,10 @@ import java.util.Arrays;
 
 import static java.nio.charset.StandardCharsets.US_ASCII;
 import static java.nio.charset.StandardCharsets.UTF_8;
-import static org.agrona.BitUtil.*;
-import static org.agrona.UnsafeAccess.UNSAFE;
-import static org.agrona.BufferUtil.*;
 import static org.agrona.AsciiEncoding.*;
+import static org.agrona.BitUtil.*;
+import static org.agrona.BufferUtil.*;
+import static org.agrona.UnsafeAccess.UNSAFE;
 
 /**
  * Expandable {@link MutableDirectBuffer} that is backed by an array. When values are put into the buffer beyond its
@@ -1263,17 +1263,21 @@ public class ExpandableArrayBuffer implements MutableDirectBuffer
 
         int start = index;
         int quotient = value;
-        int length = 1;
+        final int length;
+        int i;
         if (value < 0)
         {
             putByte0(index, MINUS_SIGN);
             start++;
-            length++;
             quotient = -quotient;
+            length = digitCount(quotient) + 1;
+            i = length - 2;
         }
-
-        int i = endOffset(quotient);
-        length += i;
+        else
+        {
+            length = digitCount(quotient);
+            i = length - 1;
+        }
 
         ensureCapacity(index, length);
 
@@ -1312,11 +1316,11 @@ public class ExpandableArrayBuffer implements MutableDirectBuffer
             return 1;
         }
 
-        int i = endOffset(value);
-        final int length = i + 1;
+        final int length = digitCount(value);
 
         ensureCapacity(index, length);
 
+        int i = length - 1;
         int quotient = value;
         final byte[] dest = byteArray;
         while (quotient >= 100)
@@ -1391,11 +1395,11 @@ public class ExpandableArrayBuffer implements MutableDirectBuffer
             return 1;
         }
 
-        int i = endOffset(value);
-        final int length = i + 1;
+        final int length = digitCount(value);
 
         ensureCapacity(index, length);
 
+        int i = length - 1;
         long quotient = value;
         final byte[] dest = byteArray;
         while (quotient >= 100000000)
@@ -1465,17 +1469,21 @@ public class ExpandableArrayBuffer implements MutableDirectBuffer
 
         int start = index;
         long quotient = value;
-        int length = 1;
+        final int length;
+        int i;
         if (value < 0)
         {
             putByte0(index, MINUS_SIGN);
             start++;
-            length++;
             quotient = -quotient;
+            length = digitCount(quotient) + 1;
+            i = length - 2;
         }
-
-        int i = endOffset(quotient);
-        length += i;
+        else
+        {
+            length = digitCount(quotient);
+            i = length - 1;
+        }
 
         ensureCapacity(index, length);