--- a/agrona/src/test/java/org/agrona/AsciiEncodingTest.java
+++ b/agrona/src/test/java/org/agrona/AsciiEncodingTest.java
@@ -15,17 +15,20 @@
  */
 package org.agrona;
 
+import org.junit.Ignore;
 import org.junit.jupiter.api.Test;
 
-import static org.agrona.AsciiEncoding.parseIntAscii;
-import static org.agrona.AsciiEncoding.parseLongAscii;
+import java.math.BigInteger;
+import java.util.stream.IntStream;
+
+import static org.agrona.AsciiEncoding.*;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-public class AsciiEncodingTest
+class AsciiEncodingTest
 {
     @Test
-    public void shouldParseInt()
+    void shouldParseInt()
     {
         assertEquals(0, parseIntAscii("0", 0, 1));
         assertEquals(0, parseIntAscii("-0", 0, 2));
@@ -48,7 +51,7 @@ public class AsciiEncodingTest
     }
 
     @Test
-    public void shouldParseLong()
+    void shouldParseLong()
     {
         assertEquals(0L, parseLongAscii("0", 0, 1));
         assertEquals(0L, parseLongAscii("-0", 0, 2));
@@ -71,7 +74,7 @@ public class AsciiEncodingTest
     }
 
     @Test
-    public void shouldThrowExceptionWhenParsingIntegerWhichCanOverFlow()
+    void shouldThrowExceptionWhenParsingIntegerWhichCanOverFlow()
     {
         final String maxValuePlusOneDigit = Integer.MAX_VALUE + "1";
         assertThrows(AsciiNumberFormatException.class,
@@ -95,7 +98,7 @@ public class AsciiEncodingTest
     }
 
     @Test
-    public void shouldThrowExceptionWhenParsingLongWhichCanOverFlow()
+    void shouldThrowExceptionWhenParsingLongWhichCanOverFlow()
     {
         final String maxValuePlusOneDigit = Long.MAX_VALUE + "1";
         assertThrows(AsciiNumberFormatException.class,
@@ -119,44 +122,114 @@ public class AsciiEncodingTest
     }
 
     @Test
-    public void shouldThrowExceptionWhenDecodingCharNonNumericValue()
+    void shouldThrowExceptionWhenDecodingCharNonNumericValue()
     {
         assertThrows(AsciiNumberFormatException.class, () -> AsciiEncoding.getDigit(0, 'a'));
     }
 
     @Test
-    public void shouldThrowExceptionWhenDecodingByteNonNumericValue()
+    void shouldThrowExceptionWhenDecodingByteNonNumericValue()
     {
         assertThrows(AsciiNumberFormatException.class, () -> AsciiEncoding.getDigit(0, (byte)'a'));
     }
 
     @Test
-    public void shouldThrowExceptionWhenParsingIntegerContainingLoneMinusSign()
+    void shouldThrowExceptionWhenParsingIntegerContainingLoneMinusSign()
     {
         assertThrows(AsciiNumberFormatException.class, () -> parseIntAscii("-", 0, 1));
     }
 
     @Test
-    public void shouldThrowExceptionWhenParsingLongContainingLoneMinusSign()
+    void shouldThrowExceptionWhenParsingLongContainingLoneMinusSign()
     {
         assertThrows(AsciiNumberFormatException.class, () -> parseLongAscii("-", 0, 1));
     }
 
     @Test
-    public void shouldThrowExceptionWhenParsingLongContainingLonePlusSign()
+    void shouldThrowExceptionWhenParsingLongContainingLonePlusSign()
     {
         assertThrows(AsciiNumberFormatException.class, () -> parseLongAscii("+", 0, 1));
     }
 
     @Test
-    public void shouldThrowExceptionWhenParsingEmptyInteger()
+    void shouldThrowExceptionWhenParsingEmptyInteger()
     {
         assertThrows(AsciiNumberFormatException.class, () -> parseIntAscii("", 0, 0));
     }
 
     @Test
-    public void shouldThrowExceptionWhenParsingEmptyLong()
+    void shouldThrowExceptionWhenParsingEmptyLong()
     {
         assertThrows(AsciiNumberFormatException.class, () -> parseLongAscii("", 0, 0));
+    }
+
+    @Test
+    void digitCountIntValue()
+    {
+        final int[] values =
+            { 9, 99, 999, 9_999, 99_999, 999_999, 9_999_999, 99_999_999, 999_999_999, Integer.MAX_VALUE };
+        for (int i = 0; i < values.length; i++)
+        {
+            assertEquals(i + 1, digitCount(values[i]));
+        }
+    }
+
+    @Test
+    void digitCountLongValue()
+    {
+        final long[] values =
+            {
+            9, 99, 999, 9_999, 99_999, 999_999, 9_999_999, 99_999_999, 999_999_999, 9_999_999_999L,
+            99_999_999_999L, 999_999_999_999L, 9_999_999_999_999L, 99_999_999_999_999L, 999999_999999999L,
+            9_999_999_999_999_999L, 99_999_999_999_999_999L, 999_999_999_999_999_999L, Long.MAX_VALUE
+            };
+        for (int i = 0; i < values.length; i++)
+        {
+            final int iter = i;
+            assertEquals(i + 1, digitCount(values[i]), () -> iter + " -> " + values[iter]);
+        }
+    }
+
+    // prints a lookup table for org.agrona.AsciiEncoding.digitCount(int)
+    @Test
+    @Ignore
+    void printDigitCountIntTable()
+    {
+        for (int i = 1; i < 33; i++)
+        {
+            final double smallest = Math.pow(2, i - 1);
+            final long log10 = (long)Math.ceil(Math.log10(smallest));
+            final long value = (long)((i < 31 ? (Math.pow(2, 32) - Math.pow(10, log10)) : 0) + (log10 << 32));
+            if (i != 1)
+            {
+                System.out.println(",");
+            }
+            System.out.print(value);
+            System.out.print("L");
+        }
+        System.out.println();
+    }
+
+    // prints a lookup table for org.agrona.AsciiEncoding.digitCount(long)
+    @Test
+    @Ignore
+    void printDigitCountLongTable()
+    {
+        final BigInteger[] pow10 = IntStream.rangeClosed(0, 19)
+            .mapToObj(BigInteger.TEN::pow)
+            .toArray(BigInteger[]::new);
+
+        for (int i = 0; i < 64; i++)
+        {
+            final int upper = i == 0 ? 0 : ((i * 1262611) >> 22) + 1;
+            final long value = ((long)(upper + 1) << 52) - pow10[upper].shiftRight(i / 4).longValue();
+            if (i != 0)
+            {
+                System.out.println(",");
+            }
+            System.out.print(value);
+            System.out.print("L");
+        }
+        System.out.println();
     }
 }