This PR based on the algorithm by Kendall Willets presented in [Computing the number of digits of an integer even faster](https://lemire.me/blog/2021/06/03/computing-the-number-of-digits-of-an-integer-even-faster/). It replaces existing `endOffset` methods with new `digitCount` methods that are branchless.

Here are the results of running `UnsafeBufferPutLongAsciiBenchmark`:
- Before
  ```
  JDK 8:
  Benchmark                                                 (value)  Mode  Cnt   Score   Error  Units
  UnsafeBufferPutLongAsciiBenchmark.benchmark  -9223372036854775808  avgt   30   5.728 ± 0.009  ns/op
  UnsafeBufferPutLongAsciiBenchmark.benchmark                     0  avgt   30   3.154 ± 0.011  ns/op
  UnsafeBufferPutLongAsciiBenchmark.benchmark                 -9182  avgt   30   6.683 ± 0.009  ns/op
  UnsafeBufferPutLongAsciiBenchmark.benchmark              97385146  avgt   30  10.186 ± 0.009  ns/op
  UnsafeBufferPutLongAsciiBenchmark.benchmark     -6180362504315475  avgt   30  18.686 ± 0.307  ns/op
  UnsafeBufferPutLongAsciiBenchmark.benchmark   9223372036854775807  avgt   30  20.517 ± 0.045  ns/op

  
  JDK 17:
  Benchmark                                                 (value)  Mode  Cnt   Score   Error  Units
  UnsafeBufferPutLongAsciiBenchmark.benchmark  -9223372036854775808  avgt   30   5.734 ± 0.010  ns/op
  UnsafeBufferPutLongAsciiBenchmark.benchmark                     0  avgt   30   2.872 ± 0.007  ns/op
  UnsafeBufferPutLongAsciiBenchmark.benchmark                 -9182  avgt   30   7.156 ± 0.006  ns/op
  UnsafeBufferPutLongAsciiBenchmark.benchmark              97385146  avgt   30  10.447 ± 0.019  ns/op
  UnsafeBufferPutLongAsciiBenchmark.benchmark     -6180362504315475  avgt   30  20.174 ± 0.060  ns/op
  UnsafeBufferPutLongAsciiBenchmark.benchmark   9223372036854775807  avgt   30  21.328 ± 0.362  ns/op
  ```
- After
  ```
  JDK 8:
  Benchmark                                                 (value)  Mode  Cnt   Score   Error  Units
  UnsafeBufferPutLongAsciiBenchmark.benchmark  -9223372036854775808  avgt   30   5.726 ± 0.006  ns/op
  UnsafeBufferPutLongAsciiBenchmark.benchmark                     0  avgt   30   3.153 ± 0.021  ns/op
  UnsafeBufferPutLongAsciiBenchmark.benchmark                 -9182  avgt   30   6.688 ± 0.010  ns/op
  UnsafeBufferPutLongAsciiBenchmark.benchmark              97385146  avgt   30  11.962 ± 0.191  ns/op
  UnsafeBufferPutLongAsciiBenchmark.benchmark     -6180362504315475  avgt   30  16.814 ± 0.217  ns/op
  UnsafeBufferPutLongAsciiBenchmark.benchmark   9223372036854775807  avgt   30  18.208 ± 0.042  ns/op

  JDK 17:
  Benchmark                                                 (value)  Mode  Cnt   Score   Error  Units
  UnsafeBufferPutLongAsciiBenchmark.benchmark  -9223372036854775808  avgt   30   5.882 ± 0.144  ns/op
  UnsafeBufferPutLongAsciiBenchmark.benchmark                     0  avgt   30   2.874 ± 0.006  ns/op
  UnsafeBufferPutLongAsciiBenchmark.benchmark                 -9182  avgt   30   6.693 ± 0.005  ns/op
  UnsafeBufferPutLongAsciiBenchmark.benchmark              97385146  avgt   30  10.388 ± 0.019  ns/op
  UnsafeBufferPutLongAsciiBenchmark.benchmark     -6180362504315475  avgt   30  17.643 ± 0.062  ns/op
  UnsafeBufferPutLongAsciiBenchmark.benchmark   9223372036854775807  avgt   30  19.767 ± 0.091  ns/op
  ```

As can be seen from the above for the big number it gives about **10%** improvement.