@mjpt777 2021-10-05
I don't see why the `-9223372036854775808` value gives better performance than shorter numbers. Is this correct?