@vyazelenko 2021-10-05
`-9223372036854775808` is `Long.MIN_VALUE` for which there is a [special case](https://github.com/real-logic/agrona/blob/66cbe36158abaf23ab15e494ae2727d59408acc3/agrona/src/main/java/org/agrona/concurrent/UnsafeBuffer.java#L2150-L2154) since it does not have a positive counter-part, i.e. `Math.abs(Long.MIN_VALUE) == Long.MIN_VALUE`. The same is true for `Integer.MAX_VALUE`.